Update Version 1.0.3
Made mods load in from a full list of all mods instead of separate lists, and then is culled by their type.
Added in some of the new 1999 weapon mods. There are a few that would require additional systems in my program.

Added in a Techrot and Scalda foe to test against.

Changed how Heavy attacks are calculated, they should be less "powerful" in melee weapons which should help with a more accurate.